+++ words/scanner_test.go
@@ -16,31 +16,26 @@ func TestScannerUnicode(t *testing.T) {
 	// From the Unicode test suite; see the gen/ folder.
 	var passed, failed int
 	for _, test := range unicodeTests {
-		test := test
-		t.Run("", func(t *testing.T) {
-			t.Parallel()
+		var scanned [][]byte
+		scanner := words.NewScanner(bytes.NewReader(test.input))
+		for scanner.Scan() {
+			scanned = append(scanned, scanner.Bytes())
+		}
 
-			var scanned [][]byte
-			scanner := words.NewScanner(bytes.NewReader(test.input))
-			for scanner.Scan() {
-				scanned = append(scanned, scanner.Bytes())
-			}
+		if err := scanner.Err(); err != nil {
+			t.Fatal(err)
+		}
 
-			if err := scanner.Err(); err != nil {
-				t.Fatal(err)
-			}
-
-			if !reflect.DeepEqual(scanned, test.expected) {
-				failed++
-				t.Errorf(`
+		if !reflect.DeepEqual(scanned, test.expected) {
+			failed++
+			t.Errorf(`
 	for input %v
 	expected  %v
 	got       %v
 	spec      %s`, test.input, test.expected, scanned, test.comment)
-			} else {
-				passed++
-			}
-		})
+		} else {
+			passed++
+		}
 	}
 	t.Logf("passed %d, failed %d", passed, failed)
 }
@@ -48,29 +43,26 @@ func TestScannerUnicode(t *testing.T) {
 // TestScannerRoundtrip tests that all input bytes are output after segmentation.
 // De facto, it also tests that we don't get infinite loops, or ever return an error.
 func TestScannerRoundtrip(t *testing.T) {
-	const runs = 2_000
+	const runs = 2000
 
 	for i := 0; i < runs; i++ {
-		t.Run("", func(t *testing.T) {
-			t.Parallel()
 
-			input := getRandomBytes()
+		input := getRandomBytes()
 
-			r := bytes.NewReader(input)
-			sc := words.NewScanner(r)
+		r := bytes.NewReader(input)
+		sc := words.NewScanner(r)
 
-			var output []byte
-			for sc.Scan() {
-				output = append(output, sc.Bytes()...)
-			}
-			if err := sc.Err(); err != nil {
-				t.Fatal(err)
-			}
+		var output []byte
+		for sc.Scan() {
+			output = append(output, sc.Bytes()...)
+		}
+		if err := sc.Err(); err != nil {
+			t.Fatal(err)
+		}
 
-			if !bytes.Equal(output, input) {
-				t.Fatal("input bytes are not the same as scanned bytes")
-			}
-		})
+		if !bytes.Equal(output, input) {
+			t.Fatal("input bytes are not the same as scanned bytes")
+		}
 	}
 }
 
@@ -97,6 +89,7 @@ func TestInvalidUTF8(t *testing.T) {
 	for sc.Scan() {
 		output = append(output, sc.Bytes()...)
 	}
+
 	if err := sc.Err(); err != nil {
 		t.Error(err)
 	}
@@ -110,20 +103,16 @@ func TestNeverZeroAtEOF(t *testing.T) {
 	// SplitFunc should never return advance = 0 when atEOF. This test is redundant
 	// with the roundtrip test above, but nice to call out this invariant.
 
-	const runs = 50
+	const runs = 100
 
 	for i := 0; i < runs; i++ {
-		t.Run("", func(t *testing.T) {
-			t.Parallel()
+		input := getRandomBytes()
 
-			input := getRandomBytes()
+		advance, _, _ := words.SplitFunc(input, true)
 
-			advance, _, _ := words.SplitFunc(input, true)
-
-			if advance == 0 {
-				t.Error("advance should never be zero when atEOF is true")
-			}
-		})
+		if advance == 0 {
+			t.Error("advance should never be zero when atEOF is true")
+		}
 	}
 }
 
@@ -131,23 +120,19 @@ func TestNeverErr(t *testing.T) {
 	// SplitFunc should never return an error. This test is redundant
 	// with the roundtrip test above, but nice to call out this invariant.
 
-	const runs = 50
+	const runs = 100
 	atEOFs := []bool{true, false}
 
 	for i := 0; i < runs; i++ {
-		t.Run("", func(t *testing.T) {
-			t.Parallel()
+		for _, atEOF := range atEOFs {
+			input := getRandomBytes()
 
-			for _, atEOF := range atEOFs {
-				input := getRandomBytes()
+			_, _, err := words.SplitFunc(input, atEOF)
 
-				_, _, err := words.SplitFunc(input, atEOF)
-
-				if err != nil {
-					t.Errorf("SplitFunc should never error (atEOF %t)", atEOF)
-				}
+			if err != nil {
+				t.Errorf("SplitFunc should never error (atEOF %t)", atEOF)
 			}
-		})
+		}
 	}
 }
